remove remainders of the arena. Auge des Drachens haben noch einige Spieler, darf nicht gelöscht werden. move item tests to tests/items.lua
