+++ src/eressea.c
@@ -8,9 +8,6 @@
 #if MUSEUM_MODULE
 #include <modules/museum.h>
 #endif
-#if ARENA_MODULE
-#include <modules/arena.h>
-#endif
 #include <triggers/triggers.h>
 #include <util/language.h>
 #include <util/functions.h>
@@ -77,9 +74,6 @@ void game_init(void)
     register_itemfunctions();
 #if MUSEUM_MODULE
     register_museum();
-#endif
-#if ARENA_MODULE
-    register_arena();
 #endif
     wormholes_register();
